Other suggested PR changes

+++ src/platform/windows/window.rs
@@ -33,7 +33,7 @@ use platform::platform::monitor::get_available_monitors;
 use platform::platform::raw_input::register_all_mice_and_keyboards_for_raw_input;
 use platform::platform::util;
 
-const WS_RESIZABLE: LONG = (winuser::WS_SIZEBOX | winuser::WS_MAXIMIZEBOX) as LONG;
+const WS_RESIZABLE: DWORD = winuser::WS_SIZEBOX | winuser::WS_MAXIMIZEBOX;
 
 /// The Win32 implementation of the main `Window` object.
 pub struct Window {
@@ -296,9 +296,9 @@ impl Window {
             winuser::GetWindowLongW(self.window.0, winuser::GWL_STYLE)
         };
         if resizable {
-            style |= WS_RESIZABLE;
+            style |= WS_RESIZABLE as LONG;
         } else {
-            style &= !WS_RESIZABLE;
+            style &= !WS_RESIZABLE as LONG;
         }
 
         unsafe {
@@ -546,9 +546,9 @@ impl Window {
             let _ = Self::grab_cursor_inner(&window, false);
 
             if resizable {
-                style |= WS_RESIZABLE;
+                style |= WS_RESIZABLE as LONG;
             } else {
-                style &= !WS_RESIZABLE;
+                style &= !WS_RESIZABLE as LONG;
             }
             winuser::SetWindowLongW(window.0, winuser::GWL_STYLE, style);
             winuser::SetWindowLongW(window.0, winuser::GWL_EXSTYLE, ex_style);
@@ -953,7 +953,7 @@ unsafe fn init(
         };
 
         if !attributes.resizable {
-            style &= !WS_RESIZABLE as u32;
+            style &= !WS_RESIZABLE;
         }
 
         if pl_attribs.parent.is_some() {
